requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.0
jupyter-client==7.3.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.1.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
terminado==0.15.0
thinc==8.0.17
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
